+++ libft.h
@@ -23,7 +23,7 @@ typedef struct s_list
 	struct s_list	*next;
 }				t_list;
 
-long	ft_atoi(const char *str);
+int		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 int		ft_isalnum(int c);
